backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.9.1
nltk==3.9.1
emoji==2.14.0
regex==2024.9.11
vaderSentiment==3.3.2
python-multipart==0.0.12
httpx==0.27.2
pydantic==2.9.2
email-validator==2.2.0
python-dotenv==1.0.1
pytest==8.3.3
